requirements.txt
python-dotenv
pandas
pymysql
sqlalchemy
tqdm
wisski_py
